+++ quilt/changelog.md
@@ -5,4 +5,5 @@
 - Fixed enter not working in the create group GUI
 - Fixed enter not working in the join group GUI
 - Added freecam support
+- Fixed microphone clipping